Add support for simple expressions

+++ model/DieFactory.c
@@ -3,7 +3,7 @@
 #include "DieFactory.h"
 
 
-Die ** diefactory_make_die_array(int sides, size_t count) {
+Die ** diefactory_make_die_array(size_t sides, size_t count) {
 	Die **dice = malloc(sizeof(Die *) * count);
 	for(size_t i = 0; i < count; ++i) {
 		dice[i] = die_create(sides);
